requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
background-tasks==0.0.1
billiard==4.2.0
boto3==1.34.5
botocore==1.34.5
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dataclasses-json==0.6.3
distro==1.8.0
Django==4.2.7
elastic-transport==8.10.0
elasticsearch==8.11.1
exceptiongroup==1.2.0
frozenlist==1.4.0
googlemaps==4.10.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
kombu==5.3.4
langchain==0.0.352
langchain-community==0.0.5
langchain-core==0.1.2
langsmith==0.0.72
marshmallow==3.20.1
mimesis==11.1.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.6.0
packaging==23.2
prompt-toolkit==3.0.41
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
random-address==1.1.1
requests==2.31.0
s3transfer==0.9.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.12
yarl==1.9.4
